+++ src/PSRule/Rules/IRuleResult.cs
@@ -1,6 +1,0 @@
-﻿namespace PSRule.Rules
-{
-    public interface IRuleResult
-    {
-    }
-}