+++ src/PSRule/Rules/Rule.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using System.ComponentModel;
 
 namespace PSRule.Rules
 {
@@ -8,11 +9,11 @@ namespace PSRule.Rules
     [JsonObject]
     public sealed class Rule
     {
+        [JsonProperty(PropertyName = "ruleId", Required = Required.Always)]
+        public string RuleId { get; set; }
+
         [JsonProperty(PropertyName = "sourcePath")]
         public string SourcePath { get; set; }
-
-        [JsonProperty(PropertyName = "name")]
-        public string Name { get; set; }
 
         [JsonProperty(PropertyName = "description")]
         public string Description { get; set; }
@@ -21,6 +22,7 @@ namespace PSRule.Rules
         /// One or more tags assigned to the rule. Tags are additional metadata used to select rules to execute and identify results.
         /// </summary>
         [JsonProperty(PropertyName = "tag")]
+        [DefaultValue(null)]
         public TagSet Tag { get; set; }
     }
 }