+++ src/PSRule/Rules/RuleFilter.cs
@@ -5,28 +5,31 @@ using System.Collections.Generic;
 namespace PSRule.Rules
 {
     /// <summary>
-    /// A filter to include or exclude rules from being processed by name or tag.
+    /// A filter to include or exclude rules from being processed by id or tag.
     /// </summary>
     public sealed class RuleFilter
     {
-        private HashSet<string> _RequiredName;
+        private HashSet<string> _RequiredRuleId;
         private Hashtable _RequiredTag;
 
         /// <summary>
-        /// Filter rules by name or tag.
+        /// Filter rules by id or tag.
         /// </summary>
-        /// <param name="name"></param>
+        /// <param name="ruleId"></param>
         /// <param name="tag"></param>
-        public RuleFilter(IEnumerable<string> name, Hashtable tag)
+        public RuleFilter(IEnumerable<string> ruleId, Hashtable tag)
         {
-            _RequiredName = name == null ? null : new HashSet<string>(name, StringComparer.OrdinalIgnoreCase);
+            _RequiredRuleId = ruleId == null ? null : new HashSet<string>(ruleId, StringComparer.OrdinalIgnoreCase);
             _RequiredTag = tag ?? null;
         }
 
-        // Matches if the Name is contained or any tag is matched
-        public bool Match(string name, TagSet tag)
+        /// <summary>
+        /// Matches if the RuleId is contained or any tag is matched
+        /// </summary>
+        /// <returns>Return true if rule is matched, otherwise false.</returns>
+        public bool Match(string ruleId, TagSet tag)
         {
-            if (_RequiredName == null || _RequiredName.Contains(name))
+            if (_RequiredRuleId == null || _RequiredRuleId.Contains(ruleId))
             {
                 if (_RequiredTag == null)
                 {
@@ -54,7 +57,7 @@ namespace PSRule.Rules
 
         public bool Match(RuleBlock block)
         {
-            return Match(block.Id, block.Tag);
+            return Match(block.RuleId, block.Tag);
         }
     }
 }