+++ tests/PSRule/PSRule.Exists.Tests.ps1
@@ -32,7 +32,7 @@ Describe 'PSRule -- Exists keyword' -Tag 'Exists' {
             $result = $testObject | Invoke-PSRule -Path (Join-Path -Path $here -ChildPath 'FromFile.Rule.ps1') -Name 'ExistsTest';
             $result | Should -Not -BeNullOrEmpty;
             $result.IsSuccess() | Should -Be $True;
-            $result.RuleName | Should -Be 'ExistsTest';
+            $result.RuleId | Should -Be 'ExistsTest';
         }
     }
 }