+++ tests/PSRule/PSRule.TypeOf.Tests.ps1
@@ -29,7 +29,7 @@ Describe 'PSRule -- TypeOf keyword' -Tag 'TypeOf' {
             $result = $testObject | Invoke-PSRule -Path (Join-Path -Path $here -ChildPath 'FromFile.Rule.ps1') -Name 'TypeOfTest';
             $result | Should -Not -BeNullOrEmpty;
             $result.IsSuccess() | Should -Be $True;
-            $result.RuleName | Should -Be 'TypeOfTest'
+            $result.RuleId | Should -Be 'TypeOfTest'
         }
     }
 }