+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 ## Unreleased
 
+- **Breaking change** - Renamed outcome filtering parameters to align to type name and increase clarity
+  - `Invoke-PSRule` has a `-Outcome` parameter instead of `-Status`
+  - `-Outcome` supports values of `Pass`, `Fail`, `Error`, `None`, `Processed` and `All`
+
 ## v0.1.0-B181222
 
 - Added rule tags to results to enable grouping and sorting [#14](https://github.com/BernieWhite/PSRule/issues/14)